Numerical approximation

We set up a container for the model parameters, but it is not exported from the module.

NumericalExample.ModelParameters — Type

A container for the model parameters. Use model_parameters to create, which performs some simple checks.

θ: weight on (log) consumption in period utility

β: discount factor

α: capital share

δ: depreciation

τ_k: tax rate on capital returns

τ_ℓ: tax rate on wages

We use a function to instantiate it, which is better style in Julia and allows for seamless extension of the interface later.

NumericalExample.model_parameters — Function

model_parameters(; θ, β, α, δ, τ_k, τ_ℓ)

Create a value that holds the parameters of the model.

Keyword arguments

θ: the parameter $θ \log(c) + (1-θ) \log(1-ℓ)$ in the utility function

β: the discount factor

α: the capital share, in the production function $k^α ℓ^{1-α}$

δ: the depreciation rate

τ_k, τ_ℓ: tax rates for capital and labor, respectively

We code the elements of the model in simple functions, which are easier to check and debug.

NumericalExample.rent_and_wage — Function

rent_and_wage(model, k, ℓ)

Calculate the rent and the wage in the model, using capital k and labor ℓ.

NumericalExample.tax_revenue — Function

tax_revenue(model, k, ℓ)

Calculate the tax revenue in the model, using capital k and labor ℓ.

NumericalExample.period_budget — Function

period_budget(model, k, ℓ)

Calculate the budget of the agent each period (can be spent on consumption and next-period capital).

NumericalExample.euler_residual — Function

euler_residual(model; c, k′, ℓ′, c′)

Calculate the residual of the intertemporal equilibrium condition (Euler equation) at a given point in time.

c is consumption for this period, k′, ℓ′, c′ are capital, labor, and consumption next period, respectively.

NumericalExample.labor_FOC_residual — Function

labor_FOC_residual(model; k, ℓ, c)

Calculate the residual of the intratemporal first order condition.

k, ℓ, c are capital, labor, and consumption this period, respectively.