Numerical approximation

Defining the model

We set up a container for the model parameters, but it is not exported from the module.

NumericalExample.ModelParameters — Type

A container for the model parameters. Use model_parameters to create, which performs some simple checks.

θ: weight on (log) consumption in period utility

β: discount factor

α: capital share

δ: depreciation

τ_k: tax rate on capital returns

τ_ℓ: tax rate on wages

We use a function to instantiate it, which is better style in Julia and allows for seamless extension of the interface later.

NumericalExample.model_parameters — Function

model_parameters(; θ, β, α, δ, τ_k, τ_ℓ)

Create a value that holds the parameters of the model.

Keyword arguments

θ: the parameter $θ \log(c) + (1-θ) \log(1-ℓ)$ in the utility function

β: the discount factor

α: the capital share, in the production function $k^α ℓ^{1-α}$

δ: the depreciation rate

τ_k, τ_ℓ: tax rates for capital and labor, respectively

Coding the relevant equilibrium conditions

We code the elements of the model in simple functions, which are easier to check and debug.

NumericalExample.rent_and_wage — Function

rent_and_wage(model, k, ℓ)

Calculate the rent and the wage in the model, using capital k and labor ℓ.

NumericalExample.tax_revenue — Function

tax_revenue(model, k, ℓ)

Calculate the tax revenue in the model, using capital k and labor ℓ.

NumericalExample.period_budget — Function

period_budget(model, k, ℓ)

Calculate the budget of the agent each period (can be spent on consumption and next-period capital).

We encapsulate equilibrium conditions into two simple functions.

NumericalExample.euler_residual — Function

euler_residual(model; c, k′, ℓ′, c′)

Calculate the residual of the intertemporal equilibrium condition (Euler equation) at a given point in time.

c is consumption for this period, k′, ℓ′, c′ are capital, labor, and consumption next period, respectively.

NumericalExample.labor_FOC_residual — Function

labor_FOC_residual(model; k, ℓ, c)

Calculate the residual of the intratemporal first order condition.

k, ℓ, c are capital, labor, and consumption this period, respectively.

Steady state calculation

Then we code the steady state as derived in the previous section.

NumericalExample.calculate_steady_state — Function

calculate_steady_state(model)

Calculate the steady state solution, returning a container with fields c̄, k̄, ℓ̄.

Spectral approximation

We solve the full model in the time domain, using Chebyshev polynomials transformed to $[0, \infty]$ (Boyd, 2001).

The method is similar to (Judd, 2002), except that we find that rational transformations of Chebyshev polynomials work well without building in exponential decay. Also, in contrast to a lot of texts, we approximate not only $k(t)$ but also $c(t)$ and $\ell(t)$, instead of deriving them from equilibrium conditions. This leads to a larger parameter space for the approximation, but ensures that the residuals are always well-defined and have finite derivatives, leading the solver quickly to the equilibrium. This requires that we transform values to be positive (for $c(t)$ and $k(t)$) and to be within $(0, 1)$ (for $\ell(t)$).

We first define the approximation setup.

NumericalExample.approximation_setup — Function

approximation_setup(
;
    basis0,
    mid_t,
    positive_transformation,
    unit_transformation
)

Return a value that contains all parameters for the approximation.

basis0: the untransformed basis (with a finite domain), defaults to a Chebyshev approximation with 10 elements, nodes on the interior of the domain

mid_t: the time period for more or less the middle node. Approximately, it should be the time horizon we are interested in, but the method is fairly accurate without determining this precisely

positive_transformation, unit_transformation: functions that transform $ℝ$ to $(0, ∞)$ and $(0, 1)$. The defaults are exp and logistic, and you probably do not need to change these.

Spectral approximations usually work best if we have a reasonably good initial guess. For this we use the steady state solution $\forall t > 0$, since that is always a solution with $k(0) = k_0$ for any initial capital level $k_0$.

NumericalExample.calculate_initial_guess — Function

calculate_initial_guess(approx, steady_state)

Calculate an initial guess for the solution from the steady state.

Solving the model

To solve the model, we define a mapping from the approximation parameters to the residuals, and use a trust region method to solve for the roots, with automatic differentiation.

The solver is encapsulated into the function below.

NumericalExample.solve_model — Function

solve_model(model, k0; approx, tol)

Solve the model using the approximation setup in approx, with initial capital k0.

tol is used for numerical tolerance.

Return an object that has fields c̃, k̃, ℓ̃ for the approximated functions on $[0,∞)$, and a boolean converged, for checking convergence.

Utilities

We made a simple function to automate plot colors and labeling.

NumericalExample.plot_vs_time — Function

plot_vs_time(fs, ts)
plot_vs_time(
    fs,
    ts,
    label;
    colors,
    graph_labels,
    label_relative_offset
)

A simple utility function to plot functions fs at times ts. graph_labels labels each graph.